(#12357) Fix root_home fact on Windows
Without this patch the root_home fact fails on windows.  This patch
fixes the problem by implementing a code block confined to windows for
the fact's value resolution.

Because there is no root account on Windows the code block simply
returns `nil` which makes the Facter fact undefined on Windows
platforms.