(maint) Update stdlib 3.x Gemfile
Without this patch we're getting build failures because our CI system is
installing dependencies directly from a git checkout of the rspec-puppet
master branch.  Recent changes to the master branch are causing failures
across the entire matrix for stdlib 3.x.

This patch addresses the problem by changing the source of truth for
dependency management.  The source of truth with this patch applied is
the Gemfile.  CI jobs may specify a specific branch of Puppet or Facter
to run against by setting the GIT_PUPPET_BRANCH or GIT_FACTER_BRANCH
environment variables respectively.